Modularize bump settings
