+++ src/commands/githubActions/bumpUpstream/settings/index.ts
@@ -1,0 +1,2 @@
+export * from "./getInitialSettings.js";
+export * from "./printSettings.js";
